Making all of the attention module to be a sub class of `BaseJAXAttentionModule`
