+++ lib/python/EasyDel/modules/roberta/modelling_roberta_flax.py
@@ -20,7 +20,8 @@ from jax.sharding import PartitionSpec
 from jax import lax, numpy as jnp
 from ..easydel_modelling_utils import EasyDelFlaxPretrainedModel
 from ..easy_attention import EasyAttention
-from ..flax_modelling_utils import get_gradient_checkpoint_policy, ACT2FN, get_dot_general_by_bits
+from ..flax_modelling_utils import get_gradient_checkpoint_policy, ACT2FN, get_dot_general_by_bits, \
+    BaseJAXAttentionModule
 
 
 class FlaxRobertaEmbeddings(nn.Module):
@@ -63,7 +64,7 @@ class FlaxRobertaEmbeddings(nn.Module):
         return hidden_states
 
 
-class FlaxRobertaSelfAttention(nn.Module):
+class FlaxRobertaSelfAttention(BaseJAXAttentionModule):
     config: RobertaConfig
     causal: bool = False
     dtype: jnp.dtype = jnp.float32
@@ -139,81 +140,6 @@ class FlaxRobertaSelfAttention(nn.Module):
 
     def _merge_heads(self, hidden_states):
         return hidden_states.reshape(hidden_states.shape[:2] + (self.config.hidden_size,))
-
-    @nn.compact
-    def _concatenate_to_cache(self, key, value, query, attention_mask):
-        """
-                The _concatenate_to_cache function is used to concatenate the key and value vectors
-                of a query with those of previous queries. This allows for the attention mechanism to
-                look at all previous queries when computing its output. The function takes in three
-                arguments: key, value, and query. It also uses two variables that are stored in the cache:
-                cached_key and cached_value.
-
-                :param self: Access the variables stored in the cache
-                :param key: Store the keys of the encoder-decoder attention
-                :param value: Initialize the cached_value variable
-                :param query: Determine the number of cache vectors to update
-                :param attention_mask: Mask out the padded vectors in the cache
-                :return: The key, value and attention_mask
-
-        """
-        is_initialized = self.has_variable("cache", "cached_key")
-        cached_key = self.variable("cache", "cached_key", jnp.zeros, key.shape, key.dtype)
-        cached_value = self.variable("cache", "cached_value", jnp.zeros, value.shape, value.dtype)
-        cache_index = self.variable("cache", "cache_index", lambda: jnp.array(0, dtype=jnp.int32))
-
-        if is_initialized:
-            *batch_dims, max_length, num_heads, depth_per_head = cached_key.value.shape
-            cur_index = cache_index.value
-            if query.shape[1] == 1:
-                mesh = self.config.jax_mesh()
-
-                def fn(
-                        _cached_key,
-                        _cached_value,
-                        _key,
-                        _value,
-                        _cur_index
-                ):
-                    assert _key.shape[1] == 1 and _value.shape[1] == 1, (_key.shape, _value.shape)
-                    sp_size = max_length // mesh.shape["sp"]
-                    axis_index = jax.lax.axis_index("sp")
-                    _cur_index = _cur_index - axis_index * sp_size
-                    _key, _value = jax.lax.cond(
-                        jnp.logical_and(cur_index >= 0, _cur_index < sp_size),
-                        lambda: (
-                            _cached_key.at[:, cur_index].set(_key[:, -1]),
-                            _cached_value.at[:, cur_index].set(_value[:, -1]),
-                        ),
-                        lambda: (_cached_key, _cached_value),
-                    )
-                    return _key, _value
-
-                fn = shard_map(
-                    fn, mesh=mesh,
-                    in_specs=(
-                        PartitionSpec(("dp", "fsdp"), "sp", "tp", None),
-                        PartitionSpec(("dp", "fsdp"), "sp", "tp", None),
-                        PartitionSpec(("dp", "fsdp"), None, "tp", None),
-                        PartitionSpec(("dp", "fsdp"), None, "tp", None),
-                        PartitionSpec()
-                    ),
-                    out_specs=(
-                        PartitionSpec(("dp", "fsdp"), "sp", "tp", None),
-                        PartitionSpec(("dp", "fsdp"), "sp", "tp", None)
-                    ),
-                    check_rep=False
-                )
-                key, value = fn(cached_key.value, cached_value.value, key, value, cur_index)
-            else:
-                indices = (0,) * len(batch_dims) + (cur_index, 0, 0)
-                key = lax.dynamic_update_slice(cached_key.value, key, indices)
-                value = lax.dynamic_update_slice(cached_value.value, value, indices)
-            cached_key.value = key
-            cached_value.value = value
-            num_updated_cache_vectors = query.shape[1]
-            cache_index.value = cache_index.value + num_updated_cache_vectors
-        return key, value, attention_mask
 
     def __call__(
             self,
